requirements.txt
numpy
PyMuPDF
rank-bm25
sentence-transformers
faiss-cpu
datasets
transformers
evaluate
scikit-learn
pandas
torch
accelerate>=0.26.0
